backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
firebase-admin==6.2.0
apscheduler==3.10.4
pytz==2024.1
vaderSentiment==3.3.2
feedparser==6.0.10
python-dateutil==2.8.2